requirements-dev.txt
certifi==2022.12.7
charset-normalizer==2.1.1
crytic-compile==0.2.4
idna==3.4
prettytable==3.5.0
pysha3==1.0.2
requests==2.28.1
slither-analyzer==0.9.1
urllib3==1.26.13
wcwidth==0.2.5
